requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==1.2.2
bleach==5.0.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
crytic-compile
cycler==0.11.0
cytoolz==0.11.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
etherscan-python==2.1.0
executing==0.8.3
fastapi==0.75.2
fastjsonschema==2.15.3
fonttools==4.33.3
frozenlist==1.3.0
h11==0.13.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
lru-dict==1.1.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
multiaddr==0.0.9
multidict==6.0.2
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netaddr==0.8.0
notebook==6.4.11
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prettytable==3.3.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
Pygments==2.12.0
pymongo==4.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
pysha3==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==23.0.0
qtconsole==5.3.0
QtPy==2.1.0
requests==2.27.1
rlp==2.0.1
scikit-learn==1.1.1
scipy==1.8.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
slither-analyzer
sniffio==1.2.0
solc-select==0.2.1
soupsieve==2.3.2.post1
stack-data==0.2.0
starlette==0.17.1
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toolz==0.11.2
tornado==6.1
traitlets==5.2.1.post0
typing_extensions==4.2.0
urllib3==1.26.8
uvicorn==0.17.6
varint==1.0.2
wcwidth==0.2.5
web3==5.28.0
webencodings==0.5.1
websockets==9.1
widgetsnbextension==3.6.0
xgboost==1.6.1
yarl==1.7.2
